Refactor git dependencies
